Refactor form for github settings
